requirements.txt
Django==5.1.3
djangorestframework==3.15.2
mongoengine==0.29.1
mongomock==4.3.0
pymongo==4.10.1
sqlparse==0.5.2
